requirements.txt
runscript
bottle
gunicorn
pymongo
future
emoji
lxml
beautifulsoup4
requests
sqlalchemy
python-telegram-bot==12.8
psycopg2-binary
feedparser
pynewtonmath
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
googletrans==3.1.0a0
google_trans_new
gpytranslate
jikanpy
speedtest-cli
coffeehouse
regex
bleach
PyYAML>=5.1.2
git+https://github.com/starry69/python-markdown2.git
wikipedia
telethon
spamwatch
alphabet_detector
pyrate-limiter
cachetools
